Move country/city set calls before longitude/latitude set calls
Changing the country and city values trigger the callback that
automatically sets the coordinates based on the city name. Thus, values
possibly entered after changing the city may update the already entered
values. The same problem occurs when a new chart is loaded with correct
coordinates but incorrect city name.